pyproject.toml
[project]
name = "ociodisplaygen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "colour-science>=0.4.6",
    "opencolorio>=2.4.1",
]
